contact endpoint protected

+++ controller/contactController.js
@@ -1,13 +1,13 @@
-const asyncHandler = require("express-async-handler");
+//const asyncHandler = require("express-async-handler");
 const Contact = require("../models/dbSchema");
-const inputValidation = require("../validation/inputValidation");
+// const inputValidation = require("../validation/inputValidation");
 /* 
 @desc Get all contacts
 @route GET /api/contacts
-@access public
+@access private
 */
 const getAllContacts = async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ userId: req.user.id });
   return res
     .status(200)
     .json({ message: "List of all contacts ", data: contacts });
@@ -15,7 +15,7 @@ const getAllContacts = async (req, res) => {
 /* 
 @desc Get a single contacts
 @route GET /api/contacts/:id
-@access public
+@access private
 */
 const getSingleContact = async (req, res) => {
   try {
@@ -33,29 +33,23 @@ const getSingleContact = async (req, res) => {
 /* 
 @desc create a single 
 @route POST /api/contacts/addContact
-@access public
+@access private
 */
 const addContact = async (req, res) => {
   const { name, email, phone } = req.body;
-  const userInput = { name, email, phone };
-  const validation = inputValidation.safeParse(userInput);
+
   try {
-    if (!validation.success) {
-      return res.status(400).json({
-        message: "validation error",
-        error: validation.error.issues[0].message,
-      });
-    }
     if (!name || !email || !phone) {
       res.status(400);
       throw new Error("All fields are required");
     }
 
-    const existUser = await Contact.findOne({ email });
-    if (existUser) {
-      return res.status(400).json({ message: "email already exist" });
-    }
-    const contact = await Contact.create({ name, email, phone });
+    const contact = await Contact.create({
+      name,
+      email,
+      phone,
+      userId: req.user.id,
+    });
     //console.log(req.body);
     return res.status(201).json({ message: "contact created", data: contact });
   } catch (error) {
@@ -66,7 +60,7 @@ const addContact = async (req, res) => {
 /* 
 @desc edit contact details
 @route PUT api/contacts/editcontact/:id
-@access public
+@access private
 */
 const editContact = async (req, res) => {
   const { name, phone } = req.body;
@@ -77,6 +71,9 @@ const editContact = async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     return res.status(404).json({ message: "contact not found" });
+  }
+  if (contact.userId.toString() !== req.user.id) {
+    return res.status(403).json({ message: "unauthorize user" });
   }
 
   const updateContact = await Contact.findByIdAndUpdate(
@@ -92,12 +89,15 @@ const editContact = async (req, res) => {
 /* 
 @desc delete a single contacts
 @route delete /api/contacts/delete/:id
-@access public
+@access private
 */
 const deleteContact = async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     return res.status(400).json("contact not found");
+  }
+  if (contact.userId.toString() !== req.user.id) {
+    return res.status(403).json({ message: "unauthorized user" });
   }
 
   await Contact.deleteOne();
